+++ ARFlight/Networking/FlightParameters.swift
@@ -1,0 +1,20 @@
+//
+//  FlightParameters.swift
+//  ARFlight
+//
+//  Created by Wandianga hassane on 22/02/2021.
+//
+
+import Foundation
+
+
+struct FlightParameters {
+    let startRange: String
+    let endRange: String
+    let departureCity: String
+    let arrivalCity: String
+    let origin: String
+    let destination: String
+    let pageSize: String
+    let pageNumber: String
+}
